# api-docs/api-docs.yaml
openapi: 3.0.3
info:
  title: PizzaFast API REST
  version: v1
  description: |
    API REST de PizzaFast para autenticacion, usuarios, catalogo, imagenes y pedidos.
    La autenticacion se basa en cookies HTTP-only generadas por los endpoints de auth.
servers:
  - url: https://localhost:8443
    description: Desarrollo local
...
paths:
...
  /api/v1/products/{productId}/images/{imageId}:
    delete:
      tags: [Products]
      summary: Eliminar imagen de producto
      security:
        - accessTokenCookie: []
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
        - $ref: '#/components/parameters/ImageIdPath'
      responses:
        '204':
          description: Imagen eliminada
        '403':
          description: Acceso denegado
  /api/v1/categories/:
    get:
      tags: [Categories]
      summary: Listar categorias
      responses:
        '200':
          description: Categorias obtenidas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDTO'
    post:
      tags: [Categories]
      summary: Crear categoria
      security:
        - accessTokenCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
      responses:
        '201':
          description: Categoria creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '403':
          description: Acceso denegado
  /api/v1/categories/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Categories]
      summary: Obtener categoria por ID
      responses:
        '200':
          description: Categoria encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '404':
          description: Categoria no encontrada
    put:
      tags: [Categories]
      summary: Actualizar categoria
      security:
        - accessTokenCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
      responses:
        '200':
          description: Categoria actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '404':
          description: Categoria no encontrada
        '403':
          description: Acceso denegado
    delete:
      tags: [Categories]
      summary: Eliminar categoria
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Categoria eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '404':
          description: Categoria no encontrada
        '403':
          description: Acceso denegado
  /api/v1/categories/{id}/products:
    get:
      tags: [Categories]
      summary: Listar productos de una categoria
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Productos obtenidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
  /api/v1/categories/{id}/images:
    post:
      tags: [Categories]
      summary: Subir imagen de categoria
      security:
        - accessTokenCookie: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [imageFile]
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        '201':
          description: Imagen creada y asociada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDTO'
        '400':
          description: Fichero invalido
        '403':
          description: Acceso denegado
  /api/v1/categories/{categoryId}/images/{imageId}:
    delete:
      tags: [Categories]
      summary: Eliminar imagen de categoria
      security:
        - accessTokenCookie: []
      parameters:
        - $ref: '#/components/parameters/CategoryIdPath'
        - $ref: '#/components/parameters/ImageIdPath'
      responses:
        '204':
          description: Imagen eliminada
        '403':
          description: Acceso denegado
...
  /api/v1/images/{id}:
    get:
      tags: [Images]
      summary: Obtener metadatos de imagen
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Imagen encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDTO'
        '404':
          description: Imagen no encontrada
  /api/v1/images/{id}/media:
    get:
      tags: [Images]
      summary: Descargar imagen
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Binario de imagen devuelto
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Imagen no encontrada
    put:
      tags: [Images]
      summary: Reemplazar imagen
      security:
        - accessTokenCookie: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [imageFile]
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        '204':
          description: Imagen reemplazada
        '401':
          description: No autenticado
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: AccessToken
      description: JWT de acceso almacenado en cookie HTTP-only.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ProductIdPath:
      name: productId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    CategoryIdPath:
      name: categoryId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ImageIdPath:
      name: imageId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
      description: Numero de pagina (base 0)
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
      description: Tamano de pagina
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Criterio de ordenacion, por ejemplo title,asc
  schemas:
...
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: integer
        allergies:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          format: int64
          nullable: true
        categoryTitle:
          type: string
          nullable: true
        hasImage:
          type: boolean
          nullable: true
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        hasImage:
          type: boolean
          nullable: true
...
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
    ProductPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductDTO'
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean